+++ src/containers/restaurant_detail.jsx
@@ -1,26 +1,15 @@
 // =============================================================
 // 
-// 	restaurant_detail.js
+// 	restaurant_detail.jsx
 //
 // =============================================================
 
-// since the book detail (meaning the area that displays the ACTIVE BOOK) makes use of this app's reducers, this is created as a 
-// CONTAINER not a component
-
 import React, { Component } from 'react';
-
-// since this component is a container, the connect function must be included
 import { connect } from 'react-redux';
-
-// imports star rating component
 import StarRatingWidget from '../components/star_rating_component';
-
-// imports review component
 import Reviews from '../components/reviews';
-
-// imports post container
 import PostReview from './restaurant_post_review';
-
+import ScrollToTop from 'react-scroll-up';
 
 class RestaurantDetail extends Component {
 
@@ -84,7 +73,7 @@ class RestaurantDetail extends Component {
 			
 			<article className="card">
 
-			    <img className="img-fluid" src="src/img/restaurant-detail.jpg" alt="Card image cap" />
+			    <img className="img-fluid" src={this.props.restaurant.image} alt={this.props.restaurant.alt} />
 
 			    <div className="card-block">
 
@@ -97,8 +86,6 @@ class RestaurantDetail extends Component {
             		</div>
 
             		<div className="detailContainer">
-
-						<img src={this.props.restaurant.image} />
 
 						<div className="starRating">
 							<StarRatingWidget rating={rating}/>
@@ -123,9 +110,11 @@ class RestaurantDetail extends Component {
 						<div className="detailBottom">
 
 							<span className="restaurantPrice">${this.props.restaurant.price} <small>avarage</small></span>
-
-						    <button className="btn btn-primary" onClick={() => this.restaurantHide()}>Return to selection</button>
-
+							<div className="return">
+							    <ScrollToTop showUnder={0}>
+									<button className="btn btn-primary" onClick={() => this.restaurantHide()}>Return to selection</button>
+								</ScrollToTop>
+							</div>
 						    <button className="btn btn-primary" onClick={event => this.postReview(event)}>Post a Review</button>
 
 						</div>